Completed part 1 & part 2
