+++ package.json
@@ -17,6 +17,7 @@
   },
   "dependencies": {
     "dotenv": "^16.4.5",
-    "express": "^4.19.2"
+    "express": "^4.19.2",
+    "test": "^3.3.0"
   }
 }
